+++ RiotNSE/NotificationService.swift
@@ -168,8 +168,8 @@ class NotificationService: UNNotificationServiceExtension {
             //  encrypted
             if !self.showDecryptedContentInNotifications {
                 //  do not show decrypted content in notification
-                NSLog("[NotificationService] fetchEvent: Do not show decrypted content in notifications.")
-                self.fallbackToBestAttemptContent(forEventId: event.eventId)
+                NSLog("[NotificationService] fetchEvent: Do not show decrypted content in notifications, no need to attempt to decrypt it.")
+                self.processEvent(event)
                 return
             }
             
@@ -399,7 +399,7 @@ class NotificationService: UNNotificationServiceExtension {
                         })
                     } else {
                         // Encrypted messages falls here
-                        notificationBody = NSString.localizedUserNotificationString(forKey: "MESSAGE_IN_X", arguments: [eventSenderName as Any])
+                        notificationBody = NSString.localizedUserNotificationString(forKey: "MESSAGE", arguments: [])
                     }
                 } else {
                     notificationTitle = eventSenderName
@@ -416,7 +416,7 @@ class NotificationService: UNNotificationServiceExtension {
                         break
                     default:
                         // Encrypted messages falls here
-                        notificationBody = NSString.localizedUserNotificationString(forKey: "MESSAGE_FROM_X", arguments: [eventSenderName as Any])
+                        notificationBody = NSString.localizedUserNotificationString(forKey: "MESSAGE", arguments: [])
                         break
                     }
                 }