+++ CHANGES.rst
@@ -39,6 +39,7 @@ Bug fix:
  * Universal links: Fix broken links for web apps (#3420).
  * SettingsViewController: Fix pan gesture crash (#3396).
  * RecentsViewController: Fix crash on dequeue some cells (#3433).
+ * NotificationService: Fix losing sound when not showing decrypted content in notifications (#3423).
 
 Changes in 0.11.6 (2020-06-30)
 ===============================================